Fix brain to handle floating point operations and add test.

+++ ELTrainingCalculator/ELTCalculatorBrain.m
@@ -69,8 +69,8 @@
 }
 
 -(double) doAction:(double (^) (double, double))action {
-    long rightArgument = [self popArgument];
-    long leftArgument = [self popArgument];
+    double rightArgument = [self popArgument];
+    double leftArgument = [self popArgument];
     double result = action(leftArgument, rightArgument);
     [self pushArgument:result];
     return result;
